22% in (note)
There are 131 underscore functions. So far, I've rewritten 29 of them (22%). The library makes good use of it's own functions to keep things clean and DRY. Since this is my first rewrite, I didn't have a plan on how best to start. My initial approach was to just grab the top file in the test spec, which was alphabetically 'Array'. 

However, after cleaning up my notes on this commit, I'm starting to see that Array's higher order of functions that rely on utility functions, which are getting in the way. I think I'm going to change up my strategy and try to find the functions that have no dependencies and work my way up. 

To be continued..